fix: Resolve GCC and Clang warnings

+++ include/huira_impl/stars/io/star_catalog.ipp
@@ -71,7 +71,7 @@ namespace huira {
         if (header.version != 1) {
             auto name = filepath.filename().string();
             HUIRA_THROW_ERROR("StarCatalog - " + name + " is out of date (version = " +
-                std::to_string(header.version) + ").  Please re-generate.");
+                std::to_string(static_cast<unsigned int>(header.version)) + ").  Please re-generate.");
         }
 
         // Binary search on file to find cutoff index
